requirements.txt
pypdf2>=3.0.0
python-docx>=1.1.0
markdown>=3.5.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
ollama>=0.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
gradio>=4.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
